Let admin can edit problems all the time.
