+++ imports/api/db.js
@@ -46,7 +46,32 @@ if (Meteor.isServer) {
     Meteor.publish('survey', function () {
         return survey.find();
     });
+    Meteor.publish('problem', function problemPublication(coding_fake) {
+        // Not using the argument passed from front-end to prevent hacking.
+        
+        var db_time = timer.findOne({timeSync: true});
+        var coding = db_time && db_time.coding;
+        var user = Meteor.users.findOne(this.userId);
+        
+        if (user && user.username && user.username == 'admin') {
+            return problem.find();
+        } else if (coding && user) {
+            return problem.find({}, {
+                fields: {
+                    description: 1,
+                    exampleInput: 1,
+                    exampleOutput: 1,
+                    images: 1,
+                    score: 1,
+                    testInput: 1,
+                    testOutput: 1,
+                    title: 1
+                }
+            });
+        } else {
+            return this.ready();
+        }
+    });
 }
-
 
 export {timer, problem, docker, userData, liveFeed, testCases, sapporo, batchAccount, requestLogs, survey};
